residuals: Combine the interface into an abstract class
